+++ api/operator/v1/vlagent_types.go
@@ -289,16 +289,13 @@ func (cr *VLAgent) DefaultStatusFields(vs *VLAgentStatus) {
 	vs.Replicas = replicaCount
 }
 
-// AnnotationsFiltered implements build.builderOpts interface
-func (cr *VLAgent) AnnotationsFiltered() map[string]string {
-	if cr.Spec.ManagedMetadata == nil {
-		return nil
+// FinalAnnotations implements build.builderOpts interface
+func (cr *VLAgent) FinalAnnotations() map[string]string {
+	var v map[string]string
+	if cr.Spec.ManagedMetadata != nil {
+		v = labels.Merge(cr.Spec.ManagedMetadata.Annotations, v)
 	}
-	dst := make(map[string]string, len(cr.Spec.ManagedMetadata.Annotations))
-	for k, v := range cr.Spec.ManagedMetadata.Annotations {
-		dst[k] = v
-	}
-	return dst
+	return v
 }
 
 // SelectorLabels returns selector labels for querying any vlagent related resources
@@ -321,14 +318,13 @@ func (cr *VLAgent) PodLabels() map[string]string {
 	return labels.Merge(cr.Spec.PodMetadata.Labels, lbls)
 }
 
-// AllLabels returns global labels for all vlagent related resources
-func (cr *VLAgent) AllLabels() map[string]string {
-	selectorLabels := cr.SelectorLabels()
-	// fast path
-	if cr.Spec.ManagedMetadata == nil {
-		return selectorLabels
+// FinalLabels returns global labels for all vlagent related resources
+func (cr *VLAgent) FinalLabels() map[string]string {
+	v := cr.SelectorLabels()
+	if cr.Spec.ManagedMetadata != nil {
+		v = labels.Merge(cr.Spec.ManagedMetadata.Labels, v)
 	}
-	return labels.Merge(selectorLabels, cr.Spec.ManagedMetadata.Labels)
+	return v
 }
 
 // PrefixedName returns name of resource with fixed prefix