+++ internal/controller/operator/factory/vtcluster/vmauth_lb.go
@@ -229,7 +229,7 @@ func buildVMauthLBDeployment(cr *vmv1.VTCluster) (*appsv1.Deployment, error) {
 func createOrUpdateVMAuthLBService(ctx context.Context, rclient client.Client, cr, prevCR *vmv1.VTCluster) error {
 	builder := func(r *vmv1.VTCluster) *build.ChildBuilder {
 		b := build.NewChildBuilder(r, vmv1beta1.ClusterComponentBalancer)
-		b.SetFinalLabels(labels.Merge(b.AllLabels(), map[string]string{
+		b.SetFinalLabels(labels.Merge(b.FinalLabels(), map[string]string{
 			vmv1beta1.VMAuthLBServiceProxyTargetLabel: "vmauth",
 		}))
 		return b
@@ -268,7 +268,7 @@ func createOrUpdatePodDisruptionBudgetForVMAuthLB(ctx context.Context, rclient c
 func createOrUpdateLBProxyService(ctx context.Context, rclient client.Client, cr, prevCR *vmv1.VTCluster, kind vmv1beta1.ClusterComponent, port, prevPort string) error {
 	builder := func(r *vmv1.VTCluster) *build.ChildBuilder {
 		b := build.NewChildBuilder(r, kind)
-		b.SetFinalLabels(labels.Merge(b.AllLabels(), map[string]string{
+		b.SetFinalLabels(labels.Merge(b.FinalLabels(), map[string]string{
 			vmv1beta1.VMAuthLBServiceProxyTargetLabel: string(kind),
 		}))
 		b.SetSelectorLabels(cr.SelectorLabels(vmv1beta1.ClusterComponentBalancer))